third_party/WebKit/LayoutTests/fast/events/touch/gesture/touch-gesture-scroll-page-past-extent-expected.txt

This tests that a gesture scroll isn't propagated from a div to the page when the div has no remaining scroll offset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
first gesture
PASS document.scrollingElement.scrollTop is 0
PASS document.scrollingElement.scrollLeft is 0
PASS wheelEventsOccurred is 0
second gesture
PASS document.scrollingElement.scrollTop is 0
PASS document.scrollingElement.scrollLeft is 0
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDivElement]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
